pyproject.toml
[project]
name = "causal-experiment"
version = "0.2.1"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "pyros-sd-models", email = "pyros.sd.models@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.2.0",
    "devtools>=0.12.2",
    "loguru>=0.7.3",
    "matplotlib>=3.10.0",
    "monsterui>=0.0.25",
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "python-fasthtml>=0.12.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers[torch]>=4.48.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }


[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[project.scripts]
causal-experiment = "causal_experiment:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
